Unify @signal. Resolves #43
